♻️ [#82] refac: 키워드 감지 및 푸시 로직 수정
